+++ apps/io-wallet-user-func/src/infra/http/handlers/__tests__/create-key-attestation.spec.ts
@@ -90,10 +90,10 @@ const keyRepository: KeyRepository = {
   getKeyByName: () =>
     TE.right(
       O.some({
-        ...privateEcKey,
+        ...publicEcKey,
         certificateChain: ["cert1", "cert2"],
         keyName: keyAttestationKeyName,
-        kid: privateEcKey.kid,
+        kid: publicEcKey.kid,
       }),
     ),
 };
@@ -457,7 +457,7 @@ describe("CreateKeyAttestationHandler", async () => {
       jose.decodeProtectedHeader(body.right.key_attestation),
     ).toMatchObject({
       alg: "ES256",
-      kid: privateEcKey.kid,
+      kid: publicEcKey.kid,
     });
   });