+++ apps/io-wallet-user-func/src/infra/http/handlers/__tests__/create-wallet-instance-attestation.spec.ts
@@ -68,10 +68,10 @@ const keyRepository: KeyRepository = {
   getKeyByName: () =>
     TE.right(
       O.some({
-        ...privateEcKey,
+        ...publicEcKey,
         certificateChain: ["cert1", "cert2"],
         keyName: walletInstanceAttestationKeyName,
-        kid: privateEcKey.kid,
+        kid: publicEcKey.kid,
       }),
     ),
 };
@@ -410,7 +410,7 @@ describe("CreateWalletInstanceAttestationHandler", async () => {
     expect(jose.decodeProtectedHeader(walletInstanceAttestation)).toMatchObject(
       {
         alg: "ES256",
-        kid: privateEcKey.kid,
+        kid: publicEcKey.kid,
       },
     );